requirements.txt
black
python-dotenv
eth-brownie==1.18.1